Merged OData 4.0 Service into master.
This merge is not done "correctly" in the sense that it is not a real git
merge, but instead a mere copy of the folders in which the ODATA4-enabled
versions of the OData Services were created in a separate branch called
"OData_4".

The problem is, inside this branch the code was created originally in the
*-odata2-* folders, then moved, then deleted, and most of these operations
were done in just a pair of commits, with the consequence that extracting
only the relevant commits by means of cherry-picking was almost impossible
(and the same for trying to perform a "git merge").